To enum all files & folders in side given Folder
This code is useful to enum all files and folders inside given folder.
This is simple version. If you find any bug please let me know to fix
it. This project is created in Visual Studion 2008 with 32 bit compiler.
This will work in unicode as well as multibyte character set.
Contact info:- mayur.shingote@yahoo.com